requirements.txt
alembic==1.14.0
bcrypt==4.0.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
numpy==1.24.3
packaging==24.2
pandas==2.1.4
plotly==5.18.0
pluggy==1.5.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.2
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
WTForms==3.2.1
